Windows: Deal with failed synchronous control transfers
A failure from a synchronous transfer (libusb0) is treated the same way
as a failure to initiate an asynchronous driver request, and the error
code (which could be ERROR_GEN_FAILURE in the case of a stall) is lost,
and LIBUSB_ERROR_IO is always returned.

Instead make the failure go through a forced sync completion, like it
does on success, and carry the error code over to the completion
handler, where it can be converted into a libusb transfer status code.
For this the overlapped Internal member is used, which is also where
async drivers are storing the error code.

The caller will then correctly receive LIBUSB_TRANSFER_STALL in case of
a stall, as would be the case for async drivers.

Fixes #1017

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>