core: add new libusb_init_context() which allows setting options during context initialization
The new initialization function addresses some shortcomings of the libusb_set_option()
interface. Namely, it allows setting the no-enumeration option (and others) on only the
contexts where it is requested. The old initialization function (libusb_init()) is
deprecated and will be removed in a future release. For now it translates to a call to
libusb_init_context() with no options specified.

Signed-off-by: Nathan Hjelm <hjelmn@google.com>