core: Tell coverity that libusb_open does not free
Internally, libusb_open does an unref in an error case. coverity doesn't
seem to notice that this is balanced with the earlier ref, and thinks
that the passed in device may be free'ed. Annotate the function to
prevent misdetections.

An alternative would be to only take the reference after checking for
the error, but the code is idiomatic as-is.

Signed-off-by: Benjamin Berg <benjamin@sipsolutions.net>