events_posix: Silence warnings about zero-allocated memory
The static analyser got confused by the fact that fds may be NULL if
there are no event sources. Obviously, in that case the later loop that
dereferences fds will never do anything, but coverity seems to miss that
part.

Silence the warning by doing an early return from the function.

Signed-off-by: Benjamin Berg <benjamin@sipsolutions.net>