linux_usbfs: Disable sleep workaround when using udev
The workaround to sleep 10ms if a device node has not yet been created
is definitely not needed with udev. I am not sure what the race looks
like in the netlink case, unless some other userspace daemon (udev) is
reacting to the same message and creates the device.

I suppose, in the long run this might be fixed by removing the netlink
code.

Signed-off-by: Benjamin Berg <benjamin@sipsolutions.net>