core: Suppress hotplug events during initial enumeration
The initial enumeration should not result in hotplug events to be fired.
This is just a convenience though, API users still need to be prepared
to be notified a second time fora device that was plugged in between
libusb_init and libusb_hotplug_register_callback.

This regressed with commit 6929b8270170 ("Fix segmentation fault in
libusb_init() if usbi_backend.init() fails"). This commit avoids the
mentioned segmentation fault by avoiding to clean up the hotplug code if
it was not yet initialised.

Closes: #1082
Related: #989