windows: Update isochronous OUT packet actual length
On Windows, unlike other platforms, the isochronous packet actual_length
value is not set on completion of OUT transfers. Our API requires the
user to check this value for isochronous transfers instead of the transfer
actual_length, if the transferred length is of interest.

The usbd Length field is not used for isochronous OUT transfers:
https://docs.microsoft.com/en-us/windows-hardware/drivers/ddi/usb/ns-usb-_usbd_iso_packet_descriptor

To make it consistent with other platforms, just return the requested
length.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>