darwin: Fix crash in log handler when stopping event thread
The darwin event thread (in contrast to other OS implementations) tries
to log to the context that created this. However, this context is only
guaranteed to be valid until the thread has started. It may be that
another context is the last one to be destroyed, causing the event
thread to log using an already destroyed context.

Fix this by only passing on ctx where it is acceptable.

Fixes: #1108