Simplify clock_gettime() conditionalization and fallback on macOS
Removed the logic in configure.ac (for Darwin only) that detected if the
clock_gettime() function exists.  Instead, just use the preprocessor to
introspect the macOS SDK version and deployment target and call
clock_gettime() if they are new enough.

Also replaced the fallback code for older macOS with
mach_absolute_time() and gettimeofday(), which are simplier because they
do not need setup and teardown carefully balanced and their usage is in
one place instead of spread in multiple places in the file.

Closes #1080