windows: Omit -mwin32 GCC option for MSYS/MinGW
Only keep it for Cygwin.

Clang doesn't support this option, and it seems only needed for
Cygwin since MinGW predefines e.g. the _WIN32 macro anyway.

From GCC documentation: -mwin32  This option is available for Cygwin
and MinGW targets. It specifies that the typical Microsoft Windows
predefined macros are to be set in the pre-processor, but does not
influence the choice of runtime library/startup code.

References #1192

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>