add unit test for libusb_set_option
The behavior of libusb_set_option was not matching the documentation whtn the
log level was set in the environment. This has been fixed but a regression test
is needed to ensure the behavior does not deviate. This commit adds a set of
unit tests to cover some of the functionality of libusb_set_option.

Signed-off-by: Nathan Hjelm <hjelmn@google.com>