darwin: add abstraction for IOUSBDeviceInterface
The IOUSBDeviceInterface object is used by libusb to interact with USB
devices macOS (Darwin, MacOS X, iOS, etc). For as long as libusb has
existed it has always used the highest available version (at compile
time) of this interface. This change breaks that by using the higest
version available between the compile environment and the version of
macOS running at execution time. This allows the same libusb build to
run on older and newer versions of macOS without disabling newer
features.

This change relies heavily on the fact that each new version of the
IOUSBDeviceInterface object builds on prior versions. This means that
libusb can use the oldest version of the interface that has a specific
function without needing to use the specific version specified at open
time which greatly simplifies the code.

Signed-off-by: Nathan Hjelm <hjelmn@google.com>