windows: No longer wait for device to get an active configuration
This commit essentially reverts commit eb164d75 (from 2018, pre 1.0.23).

Commit eb164d75 was made with the assumption that when a device is not
configured (i.e. it returns CurrentConfigurationValue == 0 during
init_device), then it is only a temporary condition caused by Windows
already listing the device but not having completed its setup yet. The
commit added a delay of up to 1 second (20 x 50ms) for re-obtaining
information and checking until CurrentConfigurationValue would
eventually change to a value > 0. The commit also forced libusb internal
structure to contain 1 as the active configuration even if the wait
failed to retrieve an active configuration, creating a discrepancy in
between the value seen from libusb and actual value managed by Windows.

There is a problem with the above because devices could be in the
unconfigured state for permanent reasons. For instance when a device is
disabled by Windows (either manually using Device Manager or
automatically because of a resource conflict). In this case the device
is listed with CurrentConfigurationValue == 0. Therefore init_device
will add a 1 second delay for each disabled device, at every device
enumeration.

A device which is not configured has no active interface nor available
endpoint. Therefore it cannot be used (except for standard request on
endpoint 0) and therefore shall not be returned by libusb as an active
device. It is the responsibility of the OS to bring the device to a
configured state. When libusb returns active configuration == 1 to a
user, the behavior of actually using this device is not defined.

The potential gain of being able to wait for Windows actually completing
device setup is not in par with the delay caused by the waiting which
has high-level impact (for instance in GUI listing devices) for standard
scenarios.

The initial problem addressed by commit eb164d75 is simply a race
condition occurring (rarely) when libusb enumeration happens exactly at
the instant when Windows is not yet done loading the driver etc. for an
enabled device. Calling init_device slightly earlier would have return
no result for this device, and slightly later would have return a fully
active device. In the first case a new enumeration would be necessary
anyway. Given hot-plug is not supported under Windows it is a good
practice to re-enumerate devices to detect newly plugged ones anyway, so
the race condition could be easily solved by simply enumerating
regularly.

To summarize: This commit reverts to handling the described race
condition by ignoring any non-configured device (not returning them from
get_device_list), instead of waiting for them to maybe get an active
configuration at a later time. The benefit is to avoid introducing
arbitrary large delay in common scenarios where some devices are
disabled at the OS level.

Moreover this commit reverts assigning a wrong value to
priv->active_config, and returning such device from get_device_list,
preventing the risk of later misuse of not configured devices.

Note: Disabled devices might not necessarily be listed under USB devices
in Windows Device Manager, but can be other system peripherals, one
example being Bluetooth chips with HCI USB interfaces.

Closes #1270