windows: Cover all enumeration passes explicitly in switch statements
This is only for readability and doesn't change the behaviour.

Several passes were implicitly grouped together in the "default" case,
which made the code unnecessarily hard to read. The reason was that the
number of EXT passes is not known, so those were conveniently treated in
the "default" case together with some non-EXT passes.

Instead introduce a "pass_type" helper variable which has a defined
range, and add all pass types explicitly in the switch statements.

At the same time update the comment which shortly explains the different
passes, it hasn't reflected reality for a long while.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>