windows: define bus numbers without using HCD
Bus numbers are defined and assigned to root hubs directly
during the hub pass of winusb libusb_get_device_list, using the
definition of a root hub as being a hub without parent at port 0.

Therefore the HCD pass is no longer used for defining root hubs.

Note: assigned bus number values for a given setup are not guaranteed
to remain constant when migrating from previous implementation to this
one.

The rational behind this change is that:

- Implementation using CM_Get_Child from HCD to get root hub devInst is
not working when using "FabulaTech Virtual USB Host Controller"
(USB-Ethernet bridge) under Hyper-V virtual machine.

- This new way of enumerating is inspired by "USB Device Viewer" Microsoft
driver sample code which is capable of correctly enumerating hubs and
devices in the above mentioned configuration as well as in regular
configurations.