Fixed a data race in log_v()
The logging function usbi_dbg() called from udev event thread needs to safely get the log level of the default context or the fallback context. This is not trivial because udev thread may log events while contexts are created and destroyed. Locking the default context mutex is not an option because usbi_dbg() may be called from code that already holds this mutex, which would lead to a deadlock. The solution implemented in this commit is to maintain a separate atomic global variable with the default debug level.