core: Avoid possible data race in log_v()
The logging function usbi_dbg() called from udev event thread needs to
safely get the log level of the default context or the fallback context.
This is not trivial because udev thread may log events while contexts
are created and destroyed. Locking the default context mutex is not an
option because usbi_dbg() may be called from code that already holds
this mutex, which would lead to a deadlock. The solution implemented in
this commit is to maintain a separate atomic global variable with the
default debug level.

The issue was found with ThreadSanitizer, although no actual bugs have
been observed.

libusb_set_debug() was also changed to call libusb_set_option() instead
of doing its own option setting which had fewer checks.

Closes #1332