tests/stress_mt: Test open/transfer/close on available devices
Verify that you can open a device, do some transfer (a device descriptor
control transfer was chosen as it should succeed with any connected
device), and close a device from multiple threads in parallel as well.

Print a warning if a device is readonly.

Consistently add thread flags to the build.

Closes #1347