core: do not use m4 to define PRINTF_FORMAT
The PRINTF_FORMAT macro is intended to allow changing how libusb gets warnings
about coding warnings for internal logging methods. At this time there is only
a single definition and it is always used. It is better to just define it in
libusbi.h and not bother with the m4.

Signed-off-by: Nathan Hjelm <hjelmn@google.com>