Fixed incorrect interpretation of COM-style Release() return value
Found thanks to a warning about implicit sign conversion.

The return value is the new reference count, not an IOReturn error code.

I'm not sure if this is copy paste, and the new ref count is unimportant, so I just preserved the existing behaviour (but adjusted the log message). The change only impacts debug logs anyway.