Fixed docs about backend functions that lock itransfer->lock
In 2013, 858b794c added a comment saying "This function gets called with the flying_transfers_lock locked"

That appears to have changed with 138b661f.  88778414 improved some stale comments, but missed these I think.

It is clear in the code that the comment is no longer true. The function is *not* called with flying_transfers_lock locked, but it does lock itransfer->lock.