Reverted seemingly harmless introduction of temporary variable
This change was made in 13a6953379fa8c94bb2476dc460439d4ba2b7c17 and this partly reverts some of it.

This change would be harmless if there was no multithreading happening, but there is.

On the hotplug background thread, darwin_devices_detached() is called, which in turn calls Release() on the device, which frees the memory, and then sets old_device->device to NULL. Shortly after, darwin_devices_attached() is called (because the kernel driver is reattaching?) and this calls darwin_get_cached_device(), which sets ->device to something new (and not NULL).

Meanwhile, back on the main thread, darwin_reenumerate_device() is running and had cached ->device in the seemingly harmless temporary variable.  But that thing was already deallocated on the other thread!

Re-reading it from the structure makes you more likely to get the value you want.

I believe there is a larger unfixed issue here, and that some mutex needs to be added somewhere. But this is a short term fix at least.