darwin: Revert seemingly harmless introduction of temporary variable
Commit 13a69533 introduced a temporary variable for the device structure
(that originally was meant to be reverted, it turns out).

Here is a follow-up to revert this part, to avoid threading issues and
reported crashes.

The temporary variable would be harmless if there was no multithreading
happening, but there is:

On the hotplug background thread, darwin_devices_detached() is called,
which in turn calls Release() on the device, which frees the memory, and
then sets old_device->device to NULL. Shortly after,
darwin_devices_attached() is called (because the kernel driver is
reattaching?) and this calls darwin_get_cached_device(), which sets
->device to something new (and not NULL).

Meanwhile, back on the main thread, darwin_reenumerate_device() is
running and had cached ->device in the seemingly harmless temporary
variable. But that thing was already deallocated on the other thread!

Re-reading it from the structure makes it more likely you get the value
you want.

There might still be unfixed multithreading issues here, but this at least
avoids an obvious regression.

Fixes #1386
Closes #1427