darwin: Avoid error checking regression
Commit 13a69533 slightly/subtly made an incorrect change to error
checking. Before that commit, we had

  if (kIOReturnSuccess != kresult || !plugInInterface) {
    return NULL;
  }

which was correct. The commit changed the function signature. Instead of
returning the pointer directly, it now returns an error code directly,
and the pointer by reference.  The above block became:

  if (kIOReturnSuccess != kresult || !plugInInterface) {
    return darwin_to_libusb(kresult);
  }

But if kresult is somehow kIOReturnSuccess but plugInInterface is NULL
(probably impossible), then we'd return LIBUSB_SUCCESS but a NULL
pointer, which is a nonsense combination.

Closes #1430