descriptor: Replace parse_descriptor() function
This function had a few problems:

- it takes two buffers as parameters but knows nothing about their length, making it easy to overrun them.
- callers make unwarranted assumptions about the alignment of structures that are passed to it (it assumes there's no padding)
- it has tricky pointer arithmetic and masking

With this new formulation, it's easier to see what's being read/written, especially the destination. It's now very clear that the destination is not being overrun because we are simply assigning to struct fields.

Also converted byte swapping macros to inline functions for more type safety.