winusb: fix winusb_get_device_list() failing to find port numbers
get_dev_port_number() was always returning port number 0 for any port, due to the incorrect assumption that strtoll's 'end' pointer will point to a null terminator on success. This was causing winusb_get_device_list() to always fail. Since we actually know the expected values of *end for both the SPDRP_LOCATION_INFORMATION and the SPDRP_LOCATION_PATHS case, check for those instead.

Additionally changed strtoll() usage to strtoul(), since the port number is unsigned and 32 bits, and changed the 'port < 0' failure condition to 'port == 0' since 0 is the actual failure return value.