examples/hotplugtest: Close old device only if new one was opened
We want to keep only one device open for simplicity, but instead of
closing the last opened device already before /trying/ to open a new
one, close it only if the new one was opened successfully. This way
we may keep the device open longer, and in the simple case where no
later arriving devices could be opened, until it is detached again.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>