descriptor: Workaround to handle device 0x04dd:0x9f51
This device has more endpoint descriptors than reported by
bNumEndpoints. The last interface does not get parsed and can't be found
since we stop parsing when we're pointing at an endpoint, but expect to
be pointing at an interface.

After parsing the endpoints, check to see if there are any left and then
skip them. This will make us point at the correct descriptor before we
continue.