introduce new transfer state LIBUSB_TRANSFER_DISCONNECT
While pollfd get a POLLERR from a fd, libusb tries to disconnect
handle from libusb. The state reports LIBUSB_TRANSFER_NO_DEVICE,
which is correct but not accurate, it is hard to distinguish from
the user side. In most cases, we get LIBUSB_TRANSFER_NO_DEVICE
through getting return value of ioctl, getting a POLLERR and
disconnect handle with a new transfer state
LIBUSB_TRANSFER_DISCONNECT is more accurate for this scenario.

Signed-off-by: maybetaken <296496296@qq.com>