Fix Zero-Length packet on Macos
Fixes issue #1581

Summary of issue:

The LIBUSB_TRANSFER_ADD_ZERO_PACKET flag does not work as expected
on Darwin. On my USB 3 device, with maxPacketSize = 1024, I see
libusb sending ZLP for packets multiple of 5120 instead of 1024.

The ZLPs are not being sent when they should because we retrieve
an incorrect value for wMaxPacketSize via GetPipePropertiesV3.

Summary of solution:
GetEndpointPropertiesV3 does retrieve unaltered wMaxPacketSize. To
call it, we leverage GetPipePropertiesV3 to set the fields required.