core: add new option to disable reap after usb is disconnected
linux:

On some Linux systems (such as Linux 6.5 and Linux 6.8),
after unplugging some USB devices, the execution time of the
usb_disconnect() function in the kernel is too long, causing
competition between the thread where handle_event() is located
and the thread in the kernel, resulting in handle_event()
thread blocking.

This option will disable Reap operation after disconnection,
even if the system has the capability. This method is used to
avoid the problem of competition between the kernel thread and
the thread where the handle_event() related function is located
after the device is unplugged, resulting in the loss of USB data
that is still online in some millisecond level communication
situations.