libusbi.h: Handle size is 0 in usbi_reallocf()
Potential of double-free in future code:
The original usbi_reallocf function does not handle the case where size
is 0 correctly. If realloc is called with a size of 0, it is equivalent
to calling free(ptr), and realloc returns NULL. If the function then
checks !ret and calls free(ptr) again, it results in a double free error,
which can lead to undefined behavior and potential crashes.

Note this can't happen with the current code, because usbi_reallocf()
is only used in one place (in Linux-only code), and size will never
be zero here.

Closes #1598