linux_udev: add workaround for udev rules
Some applications/libraries try to discover USB devices from within udev
rules. This is wrong, as the device is not initialized by udev and
should not be discovered by libusb. The libusb device discovery should
be disabled within a udev rule and libusb_wrap_sys_device should be used
instead.

Add a workaround by reading the environment variables and detecting when
we seem to be running within a udev rule as the corresponding
environment variables are set. In that case, let the application
discover the uninitialized device if it is the one we are running for.

Signed-off-by: Benjamin Berg <benjamin@sipsolutions.net>